% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burningEmbers.R
\name{burningEmbers}
\alias{burningEmbers}
\title{create \code{burningEmbers} object}
\usage{
burningEmbers(bot, top, colNames = NULL)
}
\arguments{
\item{bot}{bottom of the colour cascade, i.e. lowest value}

\item{top}{top of the colour cascade, i.e. highest value}

\item{colNames}{names of the different columns in the graph}
}
\value{
object of class \code{burningEmbers} which can be plotted.
}
\description{
Create a \code{burningEmbers} object from the given parameter.
}
\examples{
x <- burningEmbers()
plot(x)

x <- burningEmbers(
                   bot = c(1, -2, 3, -3),
                   top = c(2,  5, 4, -2)
                   )
plot(x)

}

