% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.burningEmbers.R
\name{plot.burningEmbers}
\alias{plot.burningEmbers}
\title{plot.burningEmbers}
\usage{
\method{plot}{burningEmbers}(
     x,
     ylim,
     labels,
     ny = 1000,
     relSpace = 7,
     add0line = TRUE,
     col = rev(heat.colors(1000)),
     borderB = "black",
     lwdB = 3,
     xlab = "",
     ylab = "",
     ...)
}
\arguments{
\item{x}{\code{burningEmbers} object to be plotted}

\item{ylim}{see \link{plot}}

\item{labels}{labels for the bars. If not specified, will be taken from the row names of the \code{x$colRange} parameter}

\item{ny}{default 1000; number of y blocks - 1000 should be sufficient}

\item{relSpace}{relative width of the gap between bars compared to width to bars}

\item{add0line}{default \code{TRUE}; if a line at \code{y = 0} should be drawn}

\item{col}{default \code{rev(heat.colors(1000))}; the colour to be used for the graph. The more values, the more continuous the gradient. 1000 should be fine.}

\item{borderB}{default "black"; the colour of the borders arounf=d the bars}

\item{lwdB}{default 3; the width of the borders arount=fd =he bars}

\item{xlab}{x label}

\item{ylab}{y label}

\item{...}{additional parameter for the \code{image()} function. Examples are \code{main}, ...}
}
\value{
the data structur which was plotted
}
\description{
The \code{plot} method for the \code{burningEmbers} objects
}
\details{
Plot a Burning Amber Diagram
}
\examples{
x <- burningEmbers()
plot(x)

x <- burningEmbers(
                   bot = c(1, -2, 3, -3),
                   top = c(2,  5, 4, -2)
                   )
plot(x, main = "This is a header")
plot(x, main = "This is a header", ylim = c(-1, 3))

}
\author{
Rainer M. Krug
}

